refactor(plugin-audit-log): replace hooks.ts monolith with declarative event table
PRD #274. Replaces the 679-LOC `hooks.ts` with a declarative
`auditEvents` table plus a single-loop interpreter. Each row pairs a
hook name with `subject` / `actor` / `diff` / `extra` strategies;
adding a new audit event means adding a row, not authoring a new
closure with bespoke ctx + actor + diff plumbing.

**Shape**

- `SubjectStrategy`: `extract` (calls `extractSubject`, optional payload
  transform via `from`) or `inline` (closure returns `{id, label}` for
  the four `*_meta_changed` events, `settings:group_changed`, and
  `credential:renamed` whose labels come from non-payload data).
- `ActorStrategy`: `ctx` (the common `actorOf(appCtx)` case),
  `context-actor` (`context.actor` for credential/session/api-token/
  device-code/email-change-requested), `self` (payload itself for
  sign-in/sign-out/registered, where `ctx.user` is null), and `custom`
  (one row: `user:invited`, which inspects `appCtx.user.id` to choose
  between admin-email and id-as-string labels).
- `diff?: { omit: string[] }` per row — inline-explicit. No shared
  OMIT_KEYS constants.
- `extra?: (payload, context, appCtx) => ...` — closure where variation
  is too irregular to encode declaratively.

`assertRedactionInvariants(events, SUBJECT_REQUIRED_REDACTIONS)` runs as
a test against the live table: any row whose subject type carries a
known sensitive field (today: `user.passwordHash`) must include it in
`diff.omit` or the test fails with the offending event named. Adding a
new sensitive column is a one-line edit in `SUBJECT_REQUIRED_REDACTIONS`.

**Renames**

- `hooks.ts` → `auditEvents.ts`
- `registerHooks` → `registerAuditEvents`
- `hooks.test.ts` → `auditEvents.test.ts` (git rename preserves history)

Three callers updated: `src/index.ts`, `auditEvents.test.ts`,
`auditExtension.test.ts`.

**Test layout — three layers**

1. Existing 14 behavioral tests preserved verbatim post-rename.
2. **15 backfill tests** for previously-uncovered rows (`entry:updated`,
   `entry:transition`, `entry:trashed`, `entry:meta_changed`,
   `user:deleted`, `user:meta_changed`, `user:signed_out`,
   `user:email_change_requested`, `user:email_changed`,
   `credential:revoked`, `credential:renamed`, `session:revoked`,
   `api_token:created`, `device_code:denied`, `term:created`,
   `term:meta_changed`). Each pins the row's `event` / `subjectType` /
   `subjectId` / `subjectLabel` / `properties` shape.
3. Pinned event-list (one assertion that drifts loudly on add/remove)
   plus `assertRedactionInvariants` exercised in three modes: real
   table passes, synthetic missing-passwordHash table throws naming
   the offending event, synthetic diff-less row is ignored.

**Behavioral preservation**

Every audit-row output is preserved: same `event` strings, same
`subjectType` / `subjectId` / `subjectLabel`, same `actorId` /
`actorLabel`, same `properties` envelope including `diff` field
ordering. SECURITY guarantees preserved: `inviteToken` never logged,
`passwordHash` redacted, raw settings values never recorded.

**Verification**

pnpm exec turbo run typecheck test lint format --filter
@plumix/plugin-audit-log -- 9/9 tasks, 136/136 tests pass.
pnpm exec turbo run typecheck -- 29/29 packages green.
pnpm boundaries -- no dependency violations.
pnpm knip -- clean.

Refs #274.
